Refactored school,classroom and student entities

+++ managers/entities/classroom/Classroom.manager.js
@@ -7,12 +7,28 @@ module.exports = class Clasroom {
         this.cortex              = cortex;
         this.validators          = validators; 
         this.mongomodels         = mongomodels;
-        this.httpExposed         = ['createClassroom','put=addStudentToClassrooom','delete=deleteClassroom'];
+        this.httpExposed         = ['createClassroom','put=updateClassroom','delete=deleteClassroom','get=getClassrooms','get=getClassroomStudents'];
         this.authorised          = ['school admin']
     }
 
-    async createClassroom({__longToken,__isAuthorised,name}){
-        const classroom = {name};
+    async validateClassroomAdmin(classroomId,userId){
+        const schools = await this.mongomodels.school.find({ admins: userId })
+        const schoolIds = schools.map(school => school._id);
+
+        const classroom = await this.mongomodels.Classroom.find({ schoolId: { $in: schoolIds },_id: classroomId});
+        if(!classroom){
+            return {
+                selfHandleResponse:{
+                    "ok": false,
+                    "message": "Invalid Classroom!",
+                    "code":404
+                }
+            }
+        }
+
+    }
+    async createClassroom({__longToken,__isAuthorised,name,schoolId}){
+        const classroom = {name,schoolId};
         let result = await this.validators.classroom.createClassroom(classroom);
         if(result) return result;
 
@@ -31,48 +47,79 @@ module.exports = class Clasroom {
         };
     }
 
-    async addStudentToClassrooom({__longToken,__isAuthorised,classroomId,studentId}){
-        const body = {classroomId,studentId};
-        let result = await this.validators.classroom.addStudentToClassrooom(body);
-        if(result) return result;
+    async updateClassroom({__longToken,__isAuthorised ,__validate, classroomId,name,schoolId}){
+        const userId = __longToken.userId
+        await this.validateClassroomAdmin(classroomId,userId)
 
+        const updateClassroomBody = {name,classroomId,schoolId}
 
-        const classroom = await this.mongomodels.Classroom.findById(classroomId);
-        const student =  await this.mongomodels.student.findById(studentId);
-        if (!classroom || !student) {
+        const updatedClassroom = await this.mongomodels.Classroom.findByIdAndUpdate({ _id: classroomId },updateClassroomBody,{ new: true } )
+        if(!updatedClassroom){
             return {
                 selfHandleResponse:{
                     "ok": false,
-                    "message": "Invalid Student or Classoom!",
+                    "message": "Invalid Classroom!",
                     "code":404
                 }
             }
         }
-        
-        if (classroom.students.includes(studentId)) {
-            return {
-              selfHandleResponse: {
-                "ok": false,
-                "message": "Student already exists in the Classroom!",
-                "code": 409
-              }
-            };
-          }
-        classroom.students.push(student);
-        await classroom.save() ;
-        const populatedClassroom = await classroom.populate('students')
+        const { __v,  ...classDetails } = updatedClassroom.toObject();
+
+
         return {
             selfHandleResponse:{
                 "ok": true,
-                "message": "Successfully Added Student to  Classroom ",
+                "message": "Successfully Updated Classroom ",
                 "code":200,
-                data:populatedClassroom
+                data:classDetails
+            }        
+        };
+    }
+
+    async getClassrooms({__longToken,__isAuthorised}){
+        const userId = __longToken.userId
+        const schools = await this.mongomodels.school.find({ admins: userId }).populate('admins');
+        const schoolIds = schools.map(school => school._id);
+        const classrooms = await this.mongomodels.Classroom.find({ schoolId: { $in: schoolIds } });
+        return {
+            selfHandleResponse:{
+                "ok": true,
+                "message": "",
+                "code":200,
+                data:classrooms
+            }        
+        };
+    }
+
+    async getClassroomStudents({__longToken,__isAuthorised,__validateQuery}){
+        const queryParams = __validateQuery
+        const classroom = await this.mongomodels.Classroom.findById(queryParams.classroomId);
+        const schoool = await this.mongomodels.school.find({ _id: classroom.schoolId,admins:decoded.userId })
+        if(!classroom||!schoool){
+            return  {
+                selfHandleResponse:{
+                    "ok": false,
+                    "message": "Invalid Classroom or School",
+                    "code":404
+                }
+            }
+        }
+        const students = await this.mongomodels.student.find({ classroom: queryParams.classroomId })
+        return {
+            selfHandleResponse:{
+                "ok": true,
+                "message": "",
+                "code":200,
+                data:students
             }        
         };
     }
 
     async  deleteClassroom({__longToken,__isAuthorised,__validateQuery , classroomId }) {
         const params = __validateQuery
+
+        await this.validateClassroomAdmin(params.classroomId,__longToken.userId)
+
         const result = await  this.mongomodels.Classroom.deleteOne({ _id: params.classroomId });
 
         if (result.deletedCount === 0) {
